Clean up the Fixed Time Step (#75)
* Corrected the fixed time step; entity updates should also only happen in the fixed time step. This means that the update step can easily miss inputs like jumps, so entities now have an `on_event` function that needs to be called to handle those properly.
* Disabled vsync, allowing for frame-rates much higher than 60 (my monitor refresh rate).
* Modify the renderer to only update parts of the world texture that actually changed, and stop storing a vector of pixel data as it isn't needed. This boots the game from 370fps to 470fps on my machine, so a good optimisation.
* Clean up the renderer calls and fix a bug; removed the `bind` function, and always bind the texture before drawing.
* Added buttons to the main screen for enabling/disabling vsync.
* Added `sand::format_to`, which is similar to `std::format`-like functions, but takes a buffer of chars and writes to that, returning a `std::string_view`.

+++ src/core/event.hpp
@@ -60,11 +60,11 @@ struct mouse_released_event {
 };
 
 struct mouse_moved_event {
-	glm::vec2 pos;
+	glm::ivec2 pos;
 };
 
 struct mouse_scrolled_event {
-	glm::vec2 offset;
+	glm::ivec2 offset;
 };
 
 // WINDOW EVENTS
